Add files via upload
I put together this neat little 2D fluid simulation. It's all written in C++ and, get this, it runs right on your computer's main brain using the SDL3 library. That means you won't need some super fancy graphics card for it, which is pretty sweet! Basically, it's a fun way to play around with how liquids move, and you can totally mess with it yourself.